Refactor to make some engine functionality available through util file
