+++ firmware/raspi/util.py
@@ -97,3 +97,22 @@ def sq_from_boardcell(boardcell):
 
 def uci_move_from_boardcells(source, dest):
     return sq_from_boardcell(source) + sq_from_boardcell(dest)
+
+def get_piece_info_from_square(square, grid):
+    '''Returns tuple of color and piece type from provided square.
+    '''
+    coords = get_chess_coords_from_square(square)
+    piece_w_color = grid[coords.row][coords.col]
+    if piece_w_color == '.':
+        return (None, None)
+    color = 'w' if piece_w_color.isupper() else 'b'
+    return (color, piece_w_color.lower())
+
+def get_chess_coords_from_square(square):
+    '''Converts chess square to a BoardCell.
+
+    Example: a1 <=> [0, 0], h8 <=> [7, 7], regardless of whether human plays white or black
+    pieces.
+    '''
+    # Nums correspond to row (rank), letters correspond to col (files)
+    return BoardCell(ord(square[1]) - ord('1'), ord(square[0]) - ord('a'))
